+++ src/term.c
@@ -373,6 +373,11 @@ term_reset(bool full)
     term.hide_mouse = cfg.hide_mouse;
   }
 
+  term_switch_status(false);
+  if (full) {
+    term_clear_status();
+  }
+
   if (full) {
     term.selected = false;
     term.hovering = false;
@@ -1543,6 +1548,31 @@ term_resize(int newrows, int newcols)
   term.marg_left = 0;
   term.marg_right = newcols - 1;
 
+  // Disable status area temporarily
+  // to avoid the complexity to consider it during resize or even reflow;
+  // to round up, we will finally clear it explicitly, so we also 
+  // do not need to save status area contents;
+  // ref: DEC clears status line on DECCOLM which we generalize to resize
+  int save_st_rows = term.st_rows;
+  bool save_st_act = term.st_active;
+  // adjust status line state
+  if (term.st_rows) {
+    if (save_st_act) {
+      term_switch_status(false);
+    }
+    for (int i = term.rows; i < term_allrows; i++) {
+      freeline(term.lines[i]);
+      term.lines[i] = newline(newcols, false);
+    }
+    newrows += term.st_rows;
+
+    // remove status area during resize
+    term.rows = term_allrows;
+    term.st_rows = 0;
+  }
+  // remember y of "normal" area
+  short old_term_y = term.curs.y;
+
  /*
   * Resize the screen and scrollback. We only need to shift
   * lines around within our data structures, because lineptr()
@@ -1599,7 +1629,9 @@ term_resize(int newrows, int newcols)
   // Expand the screen if newrows > rows
   if (newrows > term.rows) {
     int added = newrows - term.rows;
+    // determine how many lines to restore from scrollback
     int restore = min(added, term.tempsblines);
+    // determine how many empty lines to add
     int create = added - restore;
 
     // Fill bottom of screen with blank lines
@@ -1680,6 +1712,48 @@ term_resize(int newrows, int newcols)
   term.rows0 = newrows;
   term.cols0 = newcols;
 
+  // Status area handling
+
+  // limit status size
+  // here term.rows still include term.st_rows
+  if (save_st_rows >= term.rows / 2) {
+    save_st_rows = max(0, term.rows / 2 - 1);
+    if (!save_st_rows) {
+      // clear status mode
+      save_st_act = false;
+      term.st_type = 0;
+    }
+  }
+  // restore status geometry
+  term.rows -= save_st_rows;
+  term.st_rows = save_st_rows;
+  // fix bottom margin in case status area got resized
+  newrows = term.rows;
+  term.marg_bot = newrows - 1;
+  // restore status line area
+  if (save_st_rows) {
+    for (int i = term.rows; i < term_allrows; i++) {
+      freeline(term.lines[i]);
+      term.lines[i] = newline(newcols, false);
+    }
+    // scroll cursor position out of status line area
+    int n = old_term_y - term.rows + 1;
+    if (n > 0) {
+      // this adjustment must be done after fixing rows/st_rows
+      // but before restoring st_active
+      term_do_scroll(term.marg_top, term.marg_bot + save_st_rows, n, true);
+      curs->y = max(0, curs->y - n);
+    }
+    // restore status mode
+    // must call after restoring (and adjusting) term.rows
+    term_switch_status(save_st_act);
+  }
+  // clear status area on resize (as DEC clears on DECCOLM);
+  // also this simplifies handling above
+  if (term.st_type == 2)
+    term_clear_status();
+
+  // Return to alternate screen
   term_switch_screen(on_alt_screen, false);
 }
 
@@ -1714,6 +1788,51 @@ term_switch_screen(bool to_alt, bool reset)
 
   if (to_alt && reset)
     term_erase(false, false, true, true);
+}
+
+/*
+ * Clear status area.
+ */
+void
+term_clear_status(void)
+{
+  term_cursor_reset(&term.st_other_curs);
+  term_cursor_reset(&term.st_saved_curs);
+
+  // clear status lines
+  for (int i = term.rows; i < term_allrows; i++) {
+    freeline(term.lines[i]);
+    term.lines[i] = newline(term.cols, false);
+  }
+
+  // home status cursor position
+  if (term.st_active)
+    term.curs.y = term.rows;
+  else
+    term.st_other_curs.y = 0;  // saved status cursor is normalized to 0
+}
+
+/*
+ * Swap active status line / main display.
+ */
+void
+term_switch_status(bool status_line)
+{
+  if (status_line == term.st_active)
+    return;
+
+  term.st_active = status_line;
+
+  term_cursor oldcurs = term.curs;
+  term.curs = term.st_other_curs;
+  // saved status cursor line is normalized to 0, adjust
+  if (status_line)
+    term.curs.y += term.rows;
+  else
+    oldcurs.y -= term.rows;
+  term.st_other_curs = oldcurs;
+
+  term_update_cs();
 }
 
 /*
@@ -1838,6 +1957,13 @@ term_do_scroll(int topline, int botline, int lines, bool sb)
   if (term.hovering) {
     term.hovering = false;
     win_update(true);
+  }
+
+  // Support scrolling within (multi-line) status area
+  if (term.st_active) {
+    topline = term.rows;
+    botline = term_allrows - 1;
+    sb = false;
   }
 
   if (term.lrmargmode && (term.marg_left || term.marg_right != term.cols - 1)) {
@@ -1977,6 +2103,10 @@ clear_wrapcontd(termline * line, int y)
   }
 }
 
+// consider status area for erasing
+#define top_y (term.st_active ? term.rows : 0)
+#define bot_y (term.st_active ? term_allrows : term.rows)
+
 /*
  * Erase a large portion of the screen: the whole screen, or the
  * whole line, or parts thereof.
@@ -2002,12 +2132,12 @@ term_erase(bool selective, bool line_only, bool from_begin, bool to_end)
   }
 
   if (from_begin)
-    start = (pos){.y = line_only ? curs->y : 0, .x = 0};
+    start = (pos){.y = line_only ? curs->y : top_y, .x = 0};
   else
     start = (pos){.y = curs->y, .x = curs->x};
 
   if (to_end)
-    end = (pos){.y = line_only ? curs->y + 1 : term.rows, .x = 0};
+    end = (pos){.y = line_only ? curs->y + 1 : bot_y, .x = 0};
   else
     end = (pos){.y = curs->y, .x = curs->x}, incpos(end);
 
@@ -2026,7 +2156,7 @@ term_erase(bool selective, bool line_only, bool from_begin, bool to_end)
     * we're fully erasing them, erase by scrolling and keep the
     * lines in the scrollback. This behaviour is not compatible with xterm. */
     int scrolllines = end.y;
-    if (end.y == term.rows) {
+    if (end.y == bot_y) {
      /* Shrink until we find a non-empty row. */
       scrolllines = term_last_nonempty_line() + 1;
     }
@@ -2058,7 +2188,7 @@ term_erase(bool selective, bool line_only, bool from_begin, bool to_end)
         if (!start.x)
           clear_cc(line, -1);
       }
-      if (inclpos(start, cols) && start.y < term.rows)
+      if (inclpos(start, cols) && start.y < bot_y)
         line = term.lines[start.y];
     }
   }
@@ -2609,7 +2739,8 @@ term_paint(void)
     term.cursor_on && !term.show_other_screen
     ? term.curs.y - term.disptop : -1;
 
-  for (int i = 0; i < term.rows; i++) {
+  // Paint all lines, including status area
+  for (int i = 0; i < term_allrows; i++) {
     pos scrpos;
     scrpos.y = i + term.disptop;
     termline *line = fetch_line(scrpos.y);