+++ src/termpriv.h
@@ -26,6 +26,8 @@ extern void term_schedule_cblink(void);
 extern void term_schedule_vbell(int already_started, int startpoint);
 
 extern void term_switch_screen(bool to_alt, bool reset);
+extern void term_switch_status(bool status_line);
+extern void term_clear_status(void);
 extern void term_check_boundary(int x, int y);
 extern void term_do_scroll(int topline, int botline, int lines, bool sb);
 extern void term_erase(bool selective, bool line_only, bool from_begin, bool to_end);